% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plans_dilution_1_pd_betabinom.R
\name{compare_plans_dilution_1_pd_betabinom}
\alias{compare_plans_dilution_1_pd_betabinom}
\title{Comparison based on probability of detection curves for different dilution schemes in the first dilution stage testing.}
\usage{
compare_plans_dilution_1_pd_betabinom(S, V0, V1, Clim_lower, Clim_upper, alpha, beta)
}
\arguments{
\item{S}{amount of sample (in grams) used for diluted solution preparation}

\item{V0}{dilution volume in the first dilution stage testing.}

\item{V1}{the volume of the diluted solution used for testing (which is equal to the plated amount in this study)}

\item{Clim_lower}{an lower value of the number of colonies counted on the two dishes retained from two successive dilutions for comparison purposes.}

\item{Clim_upper}{an upper value of the number of colonies counted on the two dishes retained from two successive dilutions for comparison purposes.}

\item{alpha}{non-negative parameters of the beta distribution.}

\item{beta}{non-negative parameters of the beta distribution.}
}
\value{
Comparison based on probability of detection curves for different dilution schemes in the first dilution stage testing.
}
\description{
\code{\link{compare_plans_dilution_1_pd_betabinom}} provides graphical displays of the probability of the detection curves for dilution schemes in the first dilution stage testing.
}
\details{
\code{\link{compare_plans_dilution_1_pd_betabinom}} provides graphical displays of the probability of the detection curves for dilution schemes in the first dilution stage testing.
}
\examples{
S <- c(1,10,20)
V0 <- 100
V1 <- 1
Clim_lower <- 0
Clim_upper <- 50
alpha <- 1
beta <- 5
compare_plans_dilution_1_pd_betabinom(S, V0,V1, Clim_lower, Clim_upper, alpha, beta)
}
