% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plans_dilution_1_pd_betabinom_pois.R
\name{compare_plans_dilution_1_pd_betabinom_pois}
\alias{compare_plans_dilution_1_pd_betabinom_pois}
\title{Comparison based on beta binomial distribution-based probability of detection curves for different dilution schemes in the first dilution stage.}
\usage{
compare_plans_dilution_1_pd_betabinom_pois(S, V0, V1, lambda_lower, lambda_upper, alpha, beta)
}
\arguments{
\item{S}{amount of sample (in grams) used for diluted solution preparation}

\item{V0}{dilution volume in the first dilution stage testing.}

\item{V1}{the volume of the diluted solution used for testing (which is equal to the plated amount in this study)}

\item{lambda_lower}{an lower value of the expected number of microorganisms count per gram.}

\item{lambda_upper}{an upper value of the expected number of microorganisms count per gram.}

\item{alpha}{non-negative parameters of the beta distribution.}

\item{beta}{non-negative parameters of the beta distribution.}
}
\value{
Comparison based on probability of detection curves for different dilution schemes in the first dilution stage.
}
\description{
\code{\link{compare_plans_dilution_1_pd_betabinom_pois}} provides graphical displays of the probability of the detection curves for dilution schemes in the first dilution stage based on the beta binomial distribution, while the count of microorganisms is modelled by Poisson distribution.
}
\details{
\code{\link{compare_plans_dilution_1_pd_betabinom_pois}} provides graphical displays of the probability of the detection curves for dilution schemes in the first dilution stage based on the beta binomial distribution, while the count of microorganisms is modelled by Poisson distribution.(This section will be updated later on.)
}
\examples{
S <- c(25,125,250)
V0 <- 100
V1 <- 1
lambda_lower <- 0
lambda_upper <- 40
alpha <- 1
beta <- 5
compare_plans_dilution_1_pd_betabinom_pois(S, V0, V1, lambda_lower, lambda_upper, alpha, beta)
}
