% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plans_dilution_2_pd_binom.R
\name{compare_plans_dilution_2_pd_binom}
\alias{compare_plans_dilution_2_pd_binom}
\title{Comparison based on probability of detection curves for different dilution schemes in the first dilution stage.}
\usage{
compare_plans_dilution_2_pd_binom(S, V0, V1, V2, V3, n_sim, Clim_lower, Clim_upper)
}
\arguments{
\item{S}{amount of sample (in grams) used for diluted solution preparation}

\item{V0}{dilution volume in the first dilution stage testing.}

\item{V1}{the volume of the diluted solution used for testing (which is equal to the plated amount in this study)}

\item{V2}{dilution volume in the first dilution stage testing.}

\item{V3}{the volume of the diluted solution used for testing (which is equal to the plated amount in this study)}

\item{n_sim}{number of simulations (large simulations provide a more precise estimation).}

\item{Clim_lower}{an lower value of the number of colonies counted on the two dishes retained from two successive dilutions for comparison purposes.}

\item{Clim_upper}{an upper value of the number of colonies counted on the two dishes retained from two successive dilutions for comparison purposes.}
}
\value{
Comparison based on probability of detection curves for different dilution schemes in the first dilution stage.
}
\description{
\code{\link{compare_plans_dilution_2_pd_binom}} provides graphical displays of the probability of the detection curves for dilution schemes in the first dilution stage.
}
\details{
\code{\link{compare_plans_dilution_2_pd_binom}} provides graphical displays of the probability of the detection curves for dilution schemes in the first dilution stage.
}
\examples{
S <- c(1,10,20)
V0 <- 100
V1 <- 1
V2 <- 10
V3 <- 1
Clim_lower <- 1
Clim_upper <- 1000
n_sim <- 20000
compare_plans_dilution_2_pd_binom(S, V0, V1, V2, V3, n_sim, Clim_lower, Clim_upper)
}
