% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plans_dilution_2_pd_binom_pois.R
\name{compare_plans_dilution_2_pd_binom_pois}
\alias{compare_plans_dilution_2_pd_binom_pois}
\title{Comparison based on binomial distribution-based probability of detection curves for different dilution schemes in the second dilution stage.}
\usage{
compare_plans_dilution_2_pd_binom_pois(S, V0, V1, V2, V3, n_sim, lambda_lower, lambda_upper)
}
\arguments{
\item{S}{amount of sample (in grams) used for diluted solution preparation}

\item{V0}{dilution volume in the first dilution stage testing.}

\item{V1}{the volume of the diluted solution used for testing (which is equal to the plated amount in this study)}

\item{V2}{dilution volume in the first dilution stage testing.}

\item{V3}{the volume of the diluted solution used for testing (which is equal to the plated amount in this study)}

\item{n_sim}{number of simulations (large simulations provide a more precise estimation).}

\item{lambda_lower}{an lower value of the expected number of microorganisms count per gram.}

\item{lambda_upper}{an upper value of the expected number of microorganisms count per gram.}
}
\value{
Comparison based on probability of detection curves for different dilution schemes in the second dilution stage.
}
\description{
\code{\link{compare_plans_dilution_2_pd_binom_pois}} provides graphical displays of the probability of the detection curves for dilution schemes in the second dilution stage based on the binomial distribution, while the count of microorganisms is modelled by Poisson distribution.
}
\details{
\code{\link{compare_plans_dilution_2_pd_binom_pois}} provides graphical displays of the probability of the detection curves for dilution schemes in the second dilution stage based on the binomial distribution, while the count of microorganisms is modelled by Poisson distribution. (This section will be updated later on.)
}
\examples{
S <- c(1,10,20)
V0 <- 100
V1 <- 1
V2 <- 10
V3 <- 1
n_sim <- 20000
lambda_lower <- 1
lambda_upper <- 100
compare_plans_dilution_2_pd_binom_pois(S, V0, V1, V2, V3, n_sim, lambda_lower, lambda_upper)
}
