% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dilution_pd_curve_homogeneous.R
\name{dilution_pd_curve_homogeneous}
\alias{dilution_pd_curve_homogeneous}
\title{comparison based on probability of detection curves for different dilution schemes when the diluted sample has homogeneous contaminants.}
\usage{
dilution_pd_curve_homogeneous(lambda_low, lambda_high, a, b, pf, USL, n_sim)
}
\arguments{
\item{lambda_low}{the lower value of the expected cell count (\eqn{\lambda}) for use in the graphical display's x-axis.}

\item{lambda_high}{the upper value of the expected cell count (\eqn{\lambda}) for use in the graphical display's x-axis.}

\item{a}{lower domain of the number of cell counts.}

\item{b}{upper domain of the number of cell counts.}

\item{pf}{plating factor (pf = 1/final dilution factor).}

\item{USL}{upper specification limit.}

\item{n_sim}{number of simulations (large simulations provide more precise estimation).}
}
\value{
Probability of detection curves when the diluted sample has homogeneous contaminants.
}
\description{
\code{\link{dilution_pd_curve_homogeneous}} provides the probability of detection curves when the diluted sample has homogeneous contaminants.
}
\details{
\code{\link{dilution_pd_curve_homogeneous}} provides probability of detection curves for different dilution schemes when the diluted sample has homogeneous contaminants (this section will be updated later on).
}
\examples{
lambda_low <- 0
lambda_high <- 10
a <- 0
b <- 300
pf <- 1000
USL <- c(1000, 2000)
n_sim <- 10000
dilution_pd_curve_homogeneous(lambda_low, lambda_high, a, b, pf, USL, n_sim)
}
