% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_validation_heterogeneous.R
\name{pd_validation_heterogeneous}
\alias{pd_validation_heterogeneous}
\title{Comparison based on probability of detection curves for different dilution schemes when diluted samples collected from a heterogeneous batch.}
\usage{
pd_validation_heterogeneous(mu_low,  mu_high, sd, a, b, f, u, USL, n_sim)
}
\arguments{
\item{mu_low}{the lower value of the mean microbial count (\eqn{\mu}) for use in the graphical display's x-axis (on the log scale).}

\item{mu_high}{the upper value of the mean microbial count (\eqn{\mu}) for use in the graphical display's x-axis (on the log scale).}

\item{sd}{the standard deviation of the normal distribution (on the log scale).}

\item{a}{lower domain of the number of cell counts.}

\item{b}{upper domain of the number of cell counts.}

\item{f}{final dilution factor.}

\item{u}{amount put on the plate.}

\item{USL}{upper specification limit.}

\item{n_sim}{number of simulations (large simulations provide more precise estimations).}
}
\value{
Probability of detection curves when diluted samples collected from a heterogeneous batch.
}
\description{
\code{\link{pd_validation_heterogeneous}} provides the probability of detection curves for validate the results when samples collected from a heterogeneous batch.
}
\details{
\code{\link{pd_curves_heterogeneous}} provides probability of detection curves for different dilution schemes when samples collected from a heterogeneous batch (this section will be updated later on).
}
\examples{
mu_low <- 1
mu_high <- 12
sd <- 0.2
a <- 0
b <- 300
f <- 0.01
u <- 0.1
USL <- 1000
n_sim <- 5000
pd_validation_heterogeneous(mu_low,  mu_high, sd, a, b, f, u, USL, n_sim)
}
