% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_validation_homogeneous.R
\name{pd_validation_homogeneous}
\alias{pd_validation_homogeneous}
\title{Comparison based on probability of detection curves for different dilution schemes when diluted samples collected from a homogeneous batch.}
\usage{
pd_validation_homogeneous(lambda_low, lambda_high, a, b, f, u, USL, n_sim)
}
\arguments{
\item{lambda_low}{the lower value of the expected microbial count (\eqn{\lambda}) for use in the graphical display's x-axis.}

\item{lambda_high}{the upper value of the expected microbial count (\eqn{\lambda}) for use in the graphical display's x-axis.}

\item{a}{lower domain of the number of microbial count.}

\item{b}{upper domain of the number of microbial count.}

\item{f}{final dilution factor.}

\item{u}{amount put on the plate.}

\item{USL}{upper specification limit.}

\item{n_sim}{number of simulations (large simulations provide more precise estimations).}
}
\value{
Probability of detection curves when diluted samples collected from a homogeneous batch.
}
\description{
\code{\link{pd_validation_homogeneous}} provides the probability of detection curves for validate the results when samples collected from a homogeneous batch.
}
\details{
\code{\link{pd_curves_homogeneous}} provides probability of detection curves for different dilution schemes when samples collected from a homogeneous batch (this section will be updated later on).
}
\examples{
lambda_low <- 0
lambda_high <- 5000
a <- 0
b <- 300
f <- 0.01
u <- 0.1
USL <- 1000
n_sim <- 5000
pd_validation_homogeneous(lambda_low, lambda_high, a, b, f, u, USL, n_sim = 500)
pd_validation_homogeneous(lambda_low, lambda_high, a, b, f, u, USL, n_sim = 5000)
}
