% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_acceptance_homogeneous_multiple.R
\name{prob_acceptance_homogeneous_multiple}
\alias{prob_acceptance_homogeneous_multiple}
\title{Probability of acceptance estimation for multiple dilution schemes when diluted samples are collected from a homogeneous batch.}
\usage{
prob_acceptance_homogeneous_multiple(c, lambda, a, b, f, u, USL, n, n_sim)
}
\arguments{
\item{c}{acceptance number}

\item{lambda}{the expected microbial count (\eqn{\lambda}).}

\item{a}{lower domain of the number of microbial count.}

\item{b}{upper domain of the number of microbial count.}

\item{f}{final dilution factor.}

\item{u}{amount put on the plate.}

\item{USL}{upper specification limit.}

\item{n}{number of samples which are used for inspection.}

\item{n_sim}{number of simulations (large simulations provide more precise estimations).}
}
\value{
Probability of acceptance  when diluted samples are collected from a homogeneous batch.
}
\description{
\code{\link{prob_acceptance_homogeneous_multiple}} provides a probability of acceptance for multiple dilution schemes in the original sample when samples collected from a homogeneous batch
}
\details{
\code{\link{prob_detection_homogeneous_multiple}} provides a probability of acceptance for multiple dilution schemes in the original sample when samples collected from a homogeneous batch (this section will be updated later on).
}
\examples{
c <- 2
lambda <- 1000
a <- 0
b <- 300
f <- c(0.01,0.1,1)
u <- c(0.1,0.1,0.1)
USL <- 1000
n <- 5
n_sim <- 50000
prob_acceptance_homogeneous_multiple(c, lambda, a, b, f, u, USL, n, n_sim)
}
