% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_detect_dilution_1_betabinom.R
\name{prob_detect_dilution_1_betabinom}
\alias{prob_detect_dilution_1_betabinom}
\title{Probability of detection for a dilution scheme in the first dilution stage testing.}
\usage{
prob_detect_dilution_1_betabinom(S, V0, V1, C_cfu, alpha, beta)
}
\arguments{
\item{S}{amount of sample (in grams) used for diluted solution preparation}

\item{V0}{dilution volume in the first dilution stage testing.}

\item{V1}{the volume of the diluted solution used for testing (which is equal to the plated amount in this study)}

\item{C_cfu}{number of colonies counted on the two dishes retained from two successive dilutions.}

\item{alpha}{non-negative parameters of the beta distribution.}

\item{beta}{non-negative parameters of the beta distribution.}
}
\value{
Probability of detection for a dilution scheme in the first dilution stage testing.
}
\description{
\code{\link{prob_detect_dilution_1_betabinom}} provides the probability of detection for a dilution scheme in the first dilution stage testing.
}
\details{
\code{\link{prob_detect_dilution_1_betabinom}} provides the probability of detection for a dilution scheme in the first dilution stage testing.
}
\examples{
S <- 25
C_cfu  <- 25
V0 <- 100
V1 <- 1
alpha <- 1
beta <- 5
prob_detect_dilution_1_betabinom(S, V0, V1, C_cfu, alpha, beta)
}
