% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_detect_dilution_1_binom.R
\name{prob_detect_dilution_1_binom}
\alias{prob_detect_dilution_1_binom}
\title{Probability of detection for a dilution scheme in the first dilution stage testing.}
\usage{
prob_detect_dilution_1_binom(S, V0, V1, C_cfu)
}
\arguments{
\item{S}{amount of sample (in grams) used for diluted solution preparation}

\item{V0}{dilution volume in the first dilution stage testing.}

\item{V1}{the volume of the diluted solution used for testing (which is equal to the plated amount in this study)}

\item{C_cfu}{number of colonies counted on the two dishes retained from two successive dilutions.}
}
\value{
Probability of detection for a dilution scheme in the first dilution stage testing.
}
\description{
\code{\link{prob_detect_dilution_1_binom}} provides the probability of detection for a dilution scheme in the first dilution stage testing.
}
\details{
\code{\link{prob_detect_dilution_1_binom}} provides the probability of detection for a dilution scheme in the first dilution stage testing.
}
\examples{
S <- c(1,10,20)
C_cfu  <- 25
V0 <- 100
V1 <- 1
prob_detect_dilution_1_binom(S, V0, V1, C_cfu )
}
