% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_detect_dilution_1_binom_pois.R
\name{prob_detect_dilution_1_binom_pois}
\alias{prob_detect_dilution_1_binom_pois}
\title{Binomial distribution-based probability of detection for a dilution scheme in the first dilution stage testing.}
\usage{
prob_detect_dilution_1_binom_pois(S,lambda, V0, V1)
}
\arguments{
\item{S}{amount of sample (in grams) used for diluted solution preparation}

\item{lambda}{the expected number of microorganisms count per gram}

\item{V0}{dilution volume in the first dilution stage testing.}

\item{V1}{the volume of the diluted solution used for testing (which is equal to the plated amount in this study)}
}
\value{
Probability of detection for a dilution scheme in the first dilution stage testing.
}
\description{
\code{\link{prob_detect_dilution_1_binom_pois}} provides a probability of detection for a dilution scheme in the first dilution stage based on the binomial distribution, while the count of microorganisms is modelled by Poisson distribution.
}
\details{
\code{\link{prob_detect_dilution_1_binom_pois}} provides a probability of detection for a dilution scheme in the first dilution stage based on the binomial distribution, while the count of microorganisms is modelled by Poisson distribution.(This section will be updated later on.)
}
\examples{
S <- c(1,10,20)
lambda  <- 10
V0 <- 100
V1 <- 1
prob_detect_dilution_1_binom_pois(S,lambda, V0, V1)
}
