% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_detect_dilution_2_betabinom_pois.R
\name{prob_detect_dilution_2_betabinom_pois}
\alias{prob_detect_dilution_2_betabinom_pois}
\title{Beta binomial distribution-based probability of detection for a dilution scheme in the second dilution stage testing.}
\usage{
prob_detect_dilution_2_betabinom_pois(S, lambda, V0, V1, V2, V3, n_sim, alpha, beta)
}
\arguments{
\item{S}{amount of sample (in grams) used for diluted solution preparation}

\item{lambda}{the expected number of microorganisms count per gram}

\item{V0}{dilution volume in the first dilution stage testing.}

\item{V1}{the volume of the diluted solution used for testing (which is equal to the plated amount in this study)}

\item{V2}{dilution volume in the first dilution stage testing.}

\item{V3}{the volume of the diluted solution used for testing (which is equal to the plated amount in this study)}

\item{n_sim}{number of simulations (large simulations provide a more precise estimation).}

\item{alpha}{non-negative parameters of the beta distribution.}

\item{beta}{non-negative parameters of the beta distribution.}
}
\value{
Probability of detection for a dilution scheme in the second dilution stage testing.
}
\description{
\code{\link{prob_detect_dilution_2_betabinom_pois}} provides a probability of detection for a dilution scheme in the second dilution stage based on the beta binomial distribution, while the count of microorganisms is modelled by Poisson distribution.
}
\details{
\code{\link{prob_detect_dilution_2_betabinom_pois}} provides a probability of detection for a dilution scheme in the second dilution stage based on the beta binomial distribution, while the count of microorganisms is modelled by Poisson distribution.(This section will be updated later on.)
}
\examples{
S <- 20
lambda  <- 10
V0 <- 100
V1 <- 1
V2 <- 10
V3 <- 1
n_sim <- 25000
alpha <- 1
beta <- 5
prob_detect_dilution_2_betabinom_pois(S, lambda, V0, V1, V2, V3, n_sim, alpha, beta)
}
