% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_detect_dilution_2_binom.R
\name{prob_detect_dilution_2_binom}
\alias{prob_detect_dilution_2_binom}
\title{Probability of detection for a dilution scheme in the second dilution stage testing.}
\usage{
prob_detect_dilution_2_binom(S, V0,V1,V2,V3,C_cfu, n_sim)
}
\arguments{
\item{S}{amount of sample (in grams) used for diluted solution preparation}

\item{V0}{dilution volume in the first dilution stage testing.}

\item{V1}{the volume of the diluted solution used for testing (which is equal to the plated amount in this study)}

\item{V2}{dilution volume in the first dilution stage testing.}

\item{V3}{the volume of the diluted solution used for testing (which is equal to the plated amount in this study)}

\item{C_cfu}{number of colonies counted on the two dishes retained from two successive dilutions.}

\item{n_sim}{number of simulations (large simulations provide a more precise estimation).}
}
\value{
Probability of detection for a dilution scheme in the second dilution stage testing.
}
\description{
\code{\link{prob_detect_dilution_2_binom}} provides the probability of detection for a dilution scheme in the second dilution stage testing.
}
\details{
\code{\link{prob_detect_dilution_2_binom}} provides the probability of detection for a dilution scheme in the second dilution stage testing.
}
\examples{
S <- 20
C_cfu  <- 25
V0 <- 100
V1 <- 1
V2 <- 10
V3 <- 1
C_cfu  <- 25
n_sim <- 25000
prob_detect_dilution_2_binom(S, V0,V1,V2,V3,C_cfu, n_sim)
}
