% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_detect_dilution_2_binom_pln.R
\name{prob_detect_dilution_2_binom_pln}
\alias{prob_detect_dilution_2_binom_pln}
\title{Binomial distribution-based probability of detection for a dilution scheme in the second dilution stage testing.}
\usage{
prob_detect_dilution_2_binom_pln (S,mu,sd, V0, V1, V2, V3, n_sim)
}
\arguments{
\item{S}{amount of sample (in grams) used for diluted solution preparation}

\item{mu}{the mean microbial count (on the log scale).}

\item{sd}{the standard deviation of the normal distribution (on the log scale).}

\item{V0}{dilution volume in the first dilution stage testing.}

\item{V1}{the volume of the diluted solution used for testing (which is equal to the plated amount in this study)}

\item{V2}{dilution volume in the first dilution stage testing.}

\item{V3}{the volume of the diluted solution used for testing (which is equal to the plated amount in this study)}

\item{n_sim}{number of simulations (large simulations provide a more precise estimation).}
}
\value{
Probability of detection for a dilution scheme in the second dilution stage testing.
}
\description{
\code{\link{prob_detect_dilution_2_binom_pln}}  provides a probability of detection for a dilution scheme in the second dilution stage based on the binomial distribution, while the count of microorganisms is modelled by Poisson lognormal distribution.
}
\details{
\code{\link{prob_detect_dilution_2_binom_pln}} provides a probability of detection for a dilution scheme in the second dilution stage based on the binomial distribution, while the count of microorganisms is modelled by Poisson lognormal distribution.(This section will be updated later on.)
}
\examples{
S <- 20
mu  <- 25
sd <- 0.8
V0 <- 100
V1 <- 1
V2 <- 10
V3 <- 1
n_sim <- 25000
prob_detect_dilution_2_binom_pln (S,mu,sd, V0, V1, V2, V3, n_sim)
}
