% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_heterogeneous.R
\name{cv_heterogeneous}
\alias{cv_heterogeneous}
\alias{cv_heterogeneous_multiple}
\alias{cv_curves_heterogeneous}
\title{coefficient of variation estimation when diluted sample collected from a heterogeneous batch.}
\usage{
cv_heterogeneous(mu, sd, a, b, f, u, USL, n_sim)

cv_heterogeneous_multiple(mu, sd, a, b, f, u, USL, n_sim)

cv_curves_heterogeneous(mu_low, mu_high, sd, a, b, f, u, USL, n_sim)
}
\arguments{
\item{mu}{the mean microbial count (on the log scale).}

\item{sd}{the standard deviation of the normal distribution (on the log scale).}

\item{a}{lower domain of the number of cell counts.}

\item{b}{upper domain of the number of cell counts.}

\item{f}{final dilution factor.}

\item{u}{amount put on the plate.}

\item{USL}{upper specification limit.}

\item{n_sim}{number of simulations (large simulations provide a more precise estimation).}

\item{mu_low}{the lower value of the mean microbial count (\eqn{\mu}) for use in the graphical display's x-axis (on the log scale).}

\item{mu_high}{the upper value of the mean microbial count (\eqn{\mu}) for use in the graphical display's x-axis (on the log scale).}
}
\value{
coefficient of variation  when sample collected from a heterogeneous batch.
}
\description{
These functions provides coefficient of variation in the original sample when diluted samples collected from a heterogeneous batch.
}
\details{
These functions provides coefficient of variation in the original sample when diluted samples collected from a heterogeneous batch.
}
\examples{
mu_low <- -5
mu_high <- 10
sd <- 0.2
a <- 0
b <- 300
f <- c(0.01,0.1)
u <- c(0.1,0.1)
USL <- 1000
n_sim <- 500
cv_curves_heterogeneous(mu_low, mu_high, sd, a, b, f, u, USL, n_sim)
}
