% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_homogeneous.R
\name{cv_homogeneous}
\alias{cv_homogeneous}
\alias{cv_homogeneous_multiple}
\alias{cv_curves_homogeneous}
\title{coefficient of variation estimation when diluted sample collected from a homogeneous batch.}
\usage{
cv_homogeneous(lambda, a, b, f, u, USL, n_sim)

cv_homogeneous_multiple(lambda, a, b, f, u, USL, n_sim)

cv_curves_homogeneous(lambda_low, lambda_high, a, b, f, u, USL, n_sim)
}
\arguments{
\item{lambda}{the expected microbial count (\eqn{\lambda}).}

\item{a}{lower domain of the number of microbial count.}

\item{b}{upper domain of the number of microbial count.}

\item{f}{final dilution factor.}

\item{u}{amount put on the plate.}

\item{USL}{upper specification limit.}

\item{n_sim}{number of simulations (large simulations provide a more precise estimation).}

\item{lambda_low}{the lower value of the expected microbial count (\eqn{\lambda}) for use in the graphical display's x-axis.}

\item{lambda_high}{the upper value of the expected microbial count (\eqn{\lambda}) for use in the graphical display's x-axis.}
}
\value{
coefficient of variation when the diluted sample collected from a homogeneous batch.
}
\description{
These functions provides coefficient of variation in the original sample  when diluted samples collected from a homogeneous batch.
}
\details{
These functions provides coefficient of variation in the original sample  when diluted samples collected from a homogeneous batch.
}
\examples{
lambda_low <- 1000
lambda_high <- 8000
a <- 0
b <- 300
f <- c(0.01,0.1)
u <- c(0.1,0.1)
USL <- 1000
n_sim <- 50000
cv_curves_homogeneous(lambda_low, lambda_high, a, b, f, u, USL, n_sim)
}
