% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OC_curves_heterogeneous.R
\name{OC_curves_heterogeneous}
\alias{OC_curves_heterogeneous}
\title{Comparison based on OC curves for different dilution schemes when the diluted sample has heterogeneous contaminants.}
\usage{
OC_curves_heterogeneous(c, meanlog_low, meanlog_high, sdlog, a, b, FDF, USL, n, n_sim)
}
\arguments{
\item{c}{acceptance number}

\item{meanlog_low}{the lower value of the mean concentration (\eqn{\mu}) for use in the graphical display's x-axis.}

\item{meanlog_high}{the upper value of the mean concentration (\eqn{\mu}) for use in the graphical display's x-axis.}

\item{sdlog}{the standard deviation of the normal distribution (on the log scale).}

\item{a}{lower domain of the number of cell counts.}

\item{b}{upper domain of the number of cell counts.}

\item{FDF}{final dilution factor.}

\item{USL}{upper specification limit.}

\item{n}{number of samples which are used for inspection.}

\item{n_sim}{number of simulations (large simulations provide more precise estimation).}
}
\value{
OC curves when the diluted sample has heterogeneous contaminants.
}
\description{
\code{\link{OC_curves_heterogeneous}} provides the operating characteristic(OC) curves when diluted sample has heterogeneous contaminants.
}
\details{
\code{\link{OC_curves_heterogeneous}} provides OC curves for different dilution schemes when the diluted sample has heterogeneous contaminants (this section will be updated later on).
}
\examples{
c <- 0
meanlog_low <- -6
meanlog_high <- 3
sdlog <- 0.8
a <- 0
b <- 300
FDF <- 0.001
USL <- c(1000, 2000)
n <- 5
n_sim <- 1000
OC_curves_heterogeneous(c, meanlog_low, meanlog_high, sdlog, a, b, FDF, USL, n, n_sim)
}
