% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OC_curves_heterogeneous.R
\name{OC_curves_heterogeneous}
\alias{OC_curves_heterogeneous}
\title{Comparison based on OC curves for different dilution schemes when the diluted samples collected from a heterogeneous batch.}
\usage{
OC_curves_heterogeneous(c, meanlog_low, meanlog_high, sdlog, a, b, f, u, USL, n, n_sim)
}
\arguments{
\item{c}{acceptance number}

\item{meanlog_low}{the lower value of the mean concentration (\eqn{\mu}) for use in the graphical display's x-axis.}

\item{meanlog_high}{the upper value of the mean concentration (\eqn{\mu}) for use in the graphical display's x-axis.}

\item{sdlog}{the standard deviation of the normal distribution (on the log scale).}

\item{a}{lower domain of the number of cell counts.}

\item{b}{upper domain of the number of cell counts.}

\item{f}{final dilution factor.}

\item{u}{amount put on the plate.}

\item{USL}{upper specification limit.}

\item{n}{number of samples which are used for inspection.}

\item{n_sim}{number of simulations (large simulations provide more precise estimations).}
}
\value{
OC curves when samples collected from a heterogeneous batch.
}
\description{
\code{\link{OC_curves_heterogeneous}} provides the operating characteristic(OC) curves when samples collected from a heterogeneous batch.
}
\details{
\code{\link{OC_curves_heterogeneous}} provides OC curves for different dilution schemes when the diluted samples collected from a heterogeneous batch (this section will be updated later on).
}
\examples{
c <- 2
meanlog_low <- 4
meanlog_high <- 9
sdlog <- 0.2
a <- 0
b <- 300
f <- c(0.01,0.1)
u <- c(0.1,0.1)
USL <- 1000
n <- 5
n_sim <- 50000
OC_curves_heterogeneous(c, meanlog_low, meanlog_high, sdlog, a, b, f, USL, u, n, n_sim)
}
