% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OC_curves_homogeneous.R
\name{OC_curves_homogeneous}
\alias{OC_curves_homogeneous}
\title{Comparison based on OC curves for different dilution schemes when diluted samples collected from a homogeneous batch.}
\usage{
OC_curves_homogeneous(c, lambda_low, lambda_high, a, b, f, u, USL, n, n_sim)
}
\arguments{
\item{c}{acceptance number}

\item{lambda_low}{the lower value of the expected microbial count(\eqn{\lambda}) for use in the graphical display's x-axis.}

\item{lambda_high}{the upper value of the expected microbial count(\eqn{\lambda}) for use in the graphical display's x-axis.}

\item{a}{lower domain of the number of cell counts.}

\item{b}{upper domain of the number of cell counts.}

\item{f}{final dilution factor.}

\item{u}{amount put on the plate.}

\item{USL}{upper specification limit.}

\item{n}{number of samples which are used for inspection.}

\item{n_sim}{number of simulations (large simulations provide more precise estimations).}
}
\value{
OC curves when diluted samples collected from a homogeneous batch.
}
\description{
\code{\link{OC_curves_homogeneous}} provides the operating characteristic(OC) curves when diluted sample has homogeneous contaminants.
}
\details{
\code{\link{OC_curves_homogeneous}} provides OC curves for different dilution schemes when samples collected from a homogeneous batch (this section will be updated later on).
}
\examples{
c <- 2
lambda_low <- 0
lambda_high <- 5000
a <- 0
b <- 300
f <- c(0.01,0.1)
u <- c(0.1,0.1)
USL <- 1000
n <- 5
n_sim <- 50000
OC_curves_homogeneous(c, lambda_low, lambda_high, a, b, f, u, USL, n, n_sim)
}
