% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pd_curves_heterogeneous.R
\name{pd_curves_heterogeneous}
\alias{pd_curves_heterogeneous}
\title{comparison based on probability of detection curves for different dilution schemes when the diluted sample has heterogeneous contaminants.}
\usage{
pd_curves_heterogeneous(meanlog_low, meanlog_high, sdlog, a, b, FDF, USL, n_sim)
}
\arguments{
\item{meanlog_low}{the lower value of the mean concentration (\eqn{\mu}) for use in the graphical display's x-axis (on the log scale).}

\item{meanlog_high}{the upper value of the mean concentration (\eqn{\mu}) for use in the graphical display's x-axis (on the log scale).}

\item{sdlog}{the standard deviation of the normal distribution (on the log scale).}

\item{a}{lower domain of the number of cell counts.}

\item{b}{upper domain of the number of cell counts.}

\item{FDF}{final dilution factor.}

\item{USL}{upper specification limit.}

\item{n_sim}{number of simulations (large simulations provide more precise estimation).}
}
\value{
Probability of detection curves when the diluted sample has heterogeneous contaminants.
}
\description{
\code{\link{pd_curves_heterogeneous}} provides the probability of detection curves when the diluted sample has heterogeneous contaminants.
}
\details{
\code{\link{pd_curves_heterogeneous}} provides probability of detection curves for different dilution schemes when the diluted sample has heterogeneous contaminants (this section will be updated later on).
}
\examples{
meanlog_low <- -5
meanlog_high <- 5
sdlog <- 0.8
a <- 0
b <- 300
FDF <- 0.001
USL <- c(1000, 2000)
n_sim <- 1000
pd_curves_heterogeneous(meanlog_low, meanlog_high, sdlog, a, b, FDF, USL, n_sim)
}
