% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_acceptance_homogeneous.R
\name{prob_acceptance_homogeneous}
\alias{prob_acceptance_homogeneous}
\title{Probability of acceptance estimation when diluted sample collected from a homogeneous batch.}
\usage{
prob_acceptance_homogeneous(c, lambda, a, b, f, u, USL, n, type, n_sim)
}
\arguments{
\item{c}{acceptance number}

\item{lambda}{the expected cell count (\eqn{\lambda}).}

\item{a}{lower domain of the number of cell counts.}

\item{b}{upper domain of the number of cell counts.}

\item{f}{final dilution factor.}

\item{u}{amount put on the plate.}

\item{USL}{upper specification limit.}

\item{n}{number of samples which are used for inspection.}

\item{type}{what type of the results you would like to consider such as "theory" or "simulation" (default "theory").}

\item{n_sim}{number of simulations (large simulations provide a more precise estimation).}
}
\value{
Probability of acceptance when the diluted sample collected from a homogeneous batch.
}
\description{
\code{\link{prob_acceptance_homogeneous}} provides a probability of acceptance in the original sample when samples collected from a homogeneous batch.
}
\details{
\code{\link{prob_detection_homogeneous}} provides a probability of acceptance when samples collected from a homogeneous batch (this section will be updated later on).
}
\examples{
c <- 2
lambda <- 2000
a <- 0
b <- 300
f <- 0.01
u <- 0.1
USL <- 1000
n <- 5
prob_acceptance_homogeneous(c, lambda, a, b, f, u, USL, n)
}
