% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_acceptance_homogeneous.R
\name{prob_acceptance_homogeneous}
\alias{prob_acceptance_homogeneous}
\title{Probability of acceptance estimation when diluted sample has homogeneous contaminations.}
\usage{
prob_acceptance_homogeneous(c, lambda, a, b, pf, USL, n, n_sim)
}
\arguments{
\item{c}{acceptance number}

\item{lambda}{the expected cell count (\eqn{\lambda}).}

\item{a}{lower domain of the number of cell counts.}

\item{b}{upper domain of the number of cell counts.}

\item{pf}{plating factor (pf = 1/final dilution factor).}

\item{USL}{upper specification limit.}

\item{n}{number of samples which are used for inspection.}

\item{n_sim}{number of simulations (large simulations provide more precise estimation).}
}
\value{
Probability of acceptance when the diluted sample has homogeneous contaminants.
}
\description{
\code{\link{prob_acceptance_homogeneous}} provides a probability of acceptance in the original sample when homogeneous contamination.
}
\details{
\code{\link{prob_detection_homogeneous}} provides a probability of acceptance when the diluted sample has homogeneous contaminants (this section will be updated later on).
}
\examples{
c <- 0
lambda <- 10
a <- 0
b <- 300
pf <- 1000
USL <- 1000
n <- 5
n_sim <- 10000
prob_acceptance_homogeneous(c, lambda, a, b, pf, USL, n, n_sim)
}
