% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_detection_heterogeneous.R
\name{prob_detection_heterogeneous}
\alias{prob_detection_heterogeneous}
\title{Probability of detection estimation when diluted sample has heterogeneous contaminations.}
\usage{
prob_detection_heterogeneous(meanlog, sdlog, a, b, FDF, USL, n_sim)
}
\arguments{
\item{meanlog}{the mean concentration (on the log scale).}

\item{sdlog}{the standard deviation of the normal distribution (on the log scale).}

\item{a}{lower domain of the number of cell counts.}

\item{b}{upper domain of the number of cell counts. lower and upper truncation points (a < x <= b).}

\item{FDF}{final dilution factor.}

\item{USL}{upper specification limit.}

\item{n_sim}{number of simulations (large simulations provide more precise estimation).}
}
\value{
Probability of detection when the diluted sample has heterogeneous contaminants.
}
\description{
\code{\link{prob_detection_heterogeneous}} provides a probability of detection in the original sample when heterogeneous contamination.
}
\details{
\code{\link{prob_detection_heterogeneous}} provides a probability of detection when the diluted sample has heterogeneous contaminants. We define the random variable \eqn{X_{i}} is the number of colonies on the \eqn{i^{th}} plate.
In practice, the acceptance for countable numbers of colonies on a plate must be between 30 and 300. Therefore, we can utilise bounded distributions to model the number of colonies on a plate. heterogeneous case, we employed truncated Poisson lognormal distribution to model (this section will be updated later on).
}
\examples{
meanlog <- 2
sdlog <- 0.8
a <- 0
b <- 300
FDF <- 0.001
USL <- 1000
n_sim <- 10000
prob_detection_heterogeneous(meanlog, sdlog, a, b, FDF, USL, n_sim)
}
