% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_detection_heterogeneous.R
\name{prob_detection_heterogeneous}
\alias{prob_detection_heterogeneous}
\title{Probability of detection estimation when diluted sample collected from a heterogeneous batch.}
\usage{
prob_detection_heterogeneous(mu, sd, a, b, f, u, USL, type, n_sim)
}
\arguments{
\item{mu}{the mean microbial count (on the log scale).}

\item{sd}{the standard deviation of the normal distribution (on the log scale).}

\item{a}{lower domain of the number of cell counts.}

\item{b}{upper domain of the number of cell counts.}

\item{f}{final dilution factor.}

\item{u}{amount put on the plate.}

\item{USL}{upper specification limit.}

\item{type}{what type of the results you would like to consider such as "theory" or "simulation" (default "theory").}

\item{n_sim}{number of simulations (large simulations provide a more precise estimation).}
}
\value{
Probability of detection when sample collected from a heterogeneous batch.
}
\description{
\code{\link{prob_detection_heterogeneous}} provides a probability of detection in the original sample when samples collected from a heterogeneous batch.
}
\details{
\code{\link{prob_detection_heterogeneous}} provides a probability of detection when the diluted sample has heterogeneous contaminants. We define the random variable \eqn{X_{i}} is the number of colonies on the \eqn{i^{th}} plate.
In practice, the acceptance for countable numbers of colonies on a plate must be between 30 and 300. Therefore, we can utilise bounded distributions to model the number of colonies on a plate. In the heterogeneous case, we employed truncated Poisson lognormal distribution to model (this section will be updated later on).
}
\examples{
mu <- 2
sd <- 0.2
a <- 0
b <- 300
f <- 0.01
u <- 0.1
USL <- 1000
prob_detection_heterogeneous(mu, sd, a, b, f, u, USL)
}
