% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_detection_homogeneous.R
\name{prob_detection_homogeneous}
\alias{prob_detection_homogeneous}
\title{Probability of detection estimation when diluted sample has homogeneous contaminations.}
\usage{
prob_detection_homogeneous(lambda, a, b, pf, USL, n_sim)
}
\arguments{
\item{lambda}{the expected cell count (\eqn{\lambda}).}

\item{a}{lower domain of the number of cell counts.}

\item{b}{upper domain of the number of cell counts.}

\item{pf}{plating factor (pf = 1/final dilution factor).}

\item{USL}{upper specification limit.}

\item{n_sim}{number of simulations (large simulations provide more precise estimation).}
}
\value{
Probability of detection when the diluted sample has homogeneous contaminants.
}
\description{
\code{\link{prob_detection_homogeneous}} provides a probability of detection in the original sample when homogeneous contamination.
}
\details{
\code{\link{prob_detection_homogeneous}} provides a probability of detection when the diluted sample has homogeneous contaminants. We define the random variable \eqn{X_{i}} is the number of colonies on the \eqn{i^{th}} plate.
In practice, the acceptance for countable numbers of colonies on a plate must be between 30 and 300. Therefore, we can utilise bounded distributions to model the number of colonies on a plate. Homogeneous case, we employed truncated Poisson distribution to model (this section will be updated later on).
}
\examples{
lambda <- 20
a <- 0
b <- 300
pf <- 1000
USL <- 1000
n_sim <- 10000
prob_detection_homogeneous(lambda, a, b, pf, USL, n_sim)
}
