% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrunpoilog.R
\name{rtrunpoilog}
\alias{rtrunpoilog}
\title{Generates random deviates from truncated poisson lognormal distribution.}
\usage{
rtrunpoilog(n, meanlog, sdlog, a, b)
}
\arguments{
\item{n}{number of observations. If length(n) > 1 then the length is taken to be the number required.}

\item{meanlog}{the mean concentration (on the log scale).}

\item{sdlog}{the standard deviation of the normal distribution (on the log scale).}

\item{a}{lower truncation points.}

\item{b}{upper truncation points (a < x <= b).}
}
\value{
rtrunpoilog generates random numbers from truncated poisson lognormal distribution.
}
\description{
\code{\link{rtrunpoilog}} provides generated random numbers from truncated poisson lognormal distribution with given parameters.
}
\details{
\code{\link{rtrunpoilog}} provides generated random numbers from truncated poisson lognormal distribution with given parameters. (this section will be updated later on).
}
\examples{
n <- 100
meanlog <-  0
sdlog <-  1
a <- 0
b <- 300
rtrunpoilog(n, meanlog, sdlog, a, b)
}
