% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_concentration_heterogeneous.R
\name{true_concentration_heterogeneous}
\alias{true_concentration_heterogeneous}
\alias{true_concentration_heterogeneous_multiple}
\alias{true_concentration_curves_heterogeneous}
\title{True concentration level estimation when diluted sample collected from a heterogeneous batch.}
\usage{
true_concentration_heterogeneous(mu, sd, a, b, f, u, USL, n_sim)

true_concentration_heterogeneous_multiple(mu, sd, a, b, f, u, USL, n_sim)

true_concentration_curves_heterogeneous(mu_low, mu_high, sd, a, b, f, u, USL, n_sim)
}
\arguments{
\item{mu}{the mean concentration (on the log scale).}

\item{sd}{the standard deviation of the normal distribution (on the log scale).}

\item{a}{lower domain of the number of cell counts.}

\item{b}{upper domain of the number of cell counts.}

\item{f}{final dilution factor.}

\item{u}{amount put on the plate.}

\item{USL}{upper specification limit.}

\item{n_sim}{number of simulations (large simulations provide a more precise estimation).}

\item{mu_low}{the lower value of the mean concentration (\eqn{\mu}) for use in the graphical display's x-axis (on the log scale).}

\item{mu_high}{the upper value of the mean concentration (\eqn{\mu}) for use in the graphical display's x-axis (on the log scale).}
}
\value{
true concentration level  when sample collected from a heterogeneous batch.
}
\description{
These functions provides true concentration level in the original sample when diluted samples collected from a heterogeneous batch.
}
\details{
Let Y be the count of microorganisms and C be the true concentration level (in counts per ml).
When diluted sample collected from heterogeneous (non-homogeneous) batch, Y can be modelled by Poisson lognormal distribution
with parameter \eqn{\mu, \sigma}. Let X be the count of microorganisms on a plate,
and it can be modelled by truncated Poisson lognormal distribution with parameters \eqn{\mu_d,\sigma, a, b}.
Also, \eqn{\lambda_d} can be written in terms of \eqn{\mu},f and u. It is given by
\deqn{\mu_d = \mu + \log(f)+ \log(u)}
And the true concentration level  is given  by
\deqn{C = \frac{X}{f*u}}
where \eqn{f} is final dilution factor and \eqn{u} is amount of diluted sample on plate. Based on the literatures, we used \eqn{\sigma = 0.2} in these dilution process; see Gonzales-Barron et al. (2013, p. 370) and Schothorst et al. (2009).
}
\examples{
mu_low <- 0
mu_high <- 10
sd <- 0.2
a <- 0
b <- 300
f <- c(0.01,0.1)
u <- c(0.1,0.1)
USL <- 1000
n_sim <- 5000
true_concentration_curves_heterogeneous(mu_low, mu_high, sd, a, b, f, u, USL, n_sim)
}
\references{
\itemize{
\item Gonzales-Barron, U.A., Pilão Cadavez, V.A., Butler, F., 2013. Statistical approaches for the design of sampling plans for microbiological monitoring of foods, in:
Mathematical and Statistical Methods in Food Science and Technology. Wiley, Chichester, UK, pp.363–384.
\item Schothorst, M. van, Zwietering, M.H., Ross, T., Buchanan, R.L., Cole, M.B., 2009. Relating microbiological criteria to food safety objectives and performance objectives. Food Control 20, \href{https://doi.org/10.1016/j.foodcont.2008.11.005}{967–979}.
}
}
