% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_concentration_homogeneous.R
\name{true_concentration_homogeneous}
\alias{true_concentration_homogeneous}
\alias{true_concentration_homogeneous_multiple}
\alias{true_concentration_curves_homogeneous}
\title{True concentration level estimation when diluted sample collected from a homogeneous batch.}
\usage{
true_concentration_homogeneous(lambda, a, b, f, u, USL, n_sim)

true_concentration_homogeneous_multiple(lambda, a, b, f, u, USL, n_sim)

true_concentration_curves_homogeneous(
  lambda_low,
  lambda_high,
  a,
  b,
  f,
  u,
  USL,
  n_sim
)
}
\arguments{
\item{lambda}{the expected microbial count (\eqn{\lambda}).}

\item{a}{lower domain of the number of microbial count.}

\item{b}{upper domain of the number of microbial count.}

\item{f}{final dilution factor.}

\item{u}{amount put on the plate.}

\item{USL}{upper specification limit.}

\item{n_sim}{number of simulations (large simulations provide a more precise estimation).}

\item{lambda_low}{the lower value of the expected microbial count (\eqn{\lambda}) for use in the graphical display's x-axis.}

\item{lambda_high}{the upper value of the expected microbial count (\eqn{\lambda}) for use in the graphical display's x-axis.}
}
\value{
true concentration level when the diluted sample collected from a homogeneous batch.
}
\description{
These functions provides true concentration level in the original sample  when diluted samples collected from a homogeneous batch.
}
\details{
Let Y be the count of microorganisms and C be the true concentration level (in counts per ml).
When diluted sample collected from homogeneous batch, Y can be modelled by Poisson distribution
with parameter \eqn{\lambda}. Let X be the count of microorganisms on a plate,
and it can be modelled by truncated Poisson distribution with parameters \eqn{\lambda_d, a, b}.
Also, \eqn{\lambda_d} can be written in terms of \eqn{\lambda},f and u. It is given by
\deqn{\lambda_d = \lambda * f *u}
And the true concentration level  is given  by
\deqn{C = \frac{X}{f*u}}
}
\examples{
lambda_low <- 0
lambda_high <- 5000
a <- 0
b <- 300
f <- c(0.01,0.1)
u <- c(0.1,0.1)
USL <- 1000
n_sim <- 50000
true_concentration_curves_homogeneous(lambda_low, lambda_high, a, b, f, u, USL, n_sim)
}
